requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
python-multipart==0.0.9
Pillow==10.4.0
numpy==1.26.4
requests==2.32.3
efficientnet_pytorch==0.7.1
onnxruntime==1.16.3


